ITC230 Week 1 App
This week I created the file structure and management for my App. On the root level I created the following files index.js and package.json. Next on the public sub-folder I named it public. The following files were created home.html and about.html the structure could change as I am not sure on the theme or purpose of my app yet. Reminder this is a working in progress app that might change as the course continues.