Add files via upload
nstall the express, express-handlebars & body-parser node modules using npm,
Convert your index.js script from the previous assignment to Express application syntax,
Update your home.html page to display a ‘Search’ form with a text field corresponding the key field in your list. For example, if your application is a book list, users could search for a book title.
When submitted, the form should post to a '/detail' route in your application. The route should respond with a page that displays:
the user’s search entry (e.g. ‘Searching for [BOOK TITLE]’)
item details
a delete link, if the item is in your data array,
a not-found message if item is not in the array (e.g. ‘[BOOK TITLE] not found’)
 When clicked, the 'delete' link should call your 'delete' route with a parameter identifying the item to delete. The route should then render a view that shows the parameter value, whether it was deleted, and the remaining # of items in your list
Both detail & delete pages should have a link back to your home page